縮み方の修正

+++ Solution/App/Player.h
@@ -29,6 +29,6 @@ private:
 	Vector _pos;
 	Vector _vec;
 	Vector _dir;
-	int _range;
+	double _range;
 };
